tests/wpt/css-tests/compositing-1_dev/html/mix-blend-mode-creates-stacking-context.htm

<!DOCTYPE html>
<html><head>
        <title>CSS Test: an element with mix-blend-mode other than normal creates a stacking context</title>
        <link href="mailto:mitica@adobe.com" rel="author" title="Mihai Tica">
        <link href="mailto:mibalan@adobe.com" rel="reviewer" title="Mihai Balan">
        <link href="https://drafts.fxtf.org/compositing-1/#mix-blend-mode" rel="help">
        <meta content="dom" name="flags">
        <meta content="Test checks that applying a blendmode other than normal
        to the element must establish a new stacking context." name="assert">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="blender"></div>
        <script type="text/javascript">
            var blendModes = ["multiply", "screen", "overlay", "darken", "lighten", "color-dodge", "color-burn",
                "hard-light", "soft-light", "difference", "exclusion", "hue", "saturation", "color", "luminosity"];

            test(function() {
                var blender = document.getElementById("blender");
                for (var i = 0; i < blendModes.length; ++i) {
                    blender.style.mixBlendMode = blendModes[i];
                    var blenderStyle = window.getComputedStyle(blender);
                    assert_true(blenderStyle.zIndex !== "auto");
                }
            });
        </script>
    

</body></html>
